Moved code to their respective files and increased error size
